refactor: Use templates to populate FormSection
Noticeable rewrite, but not as big as it seems.
Basically, by transforming the old `NameSection` components into templates, we can pass them to FormSection as props.
Instead of FormSection being a useless wrapper, we can reuse *A LOT* of logic, so the Form component is way more mantainable.
Another important change was normalising form-data. This way, PersonalSection does not have its own weird logic. We can just indicate to FormSection that this section is not extendable, and thus does not need a delete button.